+++ punchbowl/data/punch_io.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
 import os
+import string
 import hashlib
 import os.path
 import subprocess
+from typing import Any
 from pathlib import Path
 
 import astropy.units as u
@@ -44,8 +46,19 @@ def get_base_file_name(cube: NDCube) -> str:
     type_code = cube.meta["TYPECODE"].value
     date_string = cube.meta.datetime.strftime("%Y%m%d%H%M%S")
     file_version = cube.meta["FILEVRSN"].value
-    file_version = "1" if file_version == "" else file_version  # file version should never be empty!
+    file_version = "?" if file_version == "" else file_version  # file version should never be empty!
     return "PUNCH_L" + file_level + "_" + type_code + obscode + "_" + date_string + "_v" + file_version
+
+
+class DefaultFormatter(string.Formatter):
+    """A formatter that doesn't fail if a keyword is missing. Used for quicklook."""
+
+    def get_field(self, field_name: str, args: Any, kwargs: Any) -> str:
+        """Provide a special getter that returns the name if it fails."""
+        try:
+            return super().get_field(field_name, args, kwargs)
+        except (KeyError, AttributeError, IndexError):
+            return "{" + field_name + "}", ()
 
 
 def _header_to_xml(header: Header) -> et.Element:
@@ -139,7 +152,8 @@ def write_ndcube_to_quicklook(cube: NDCube,
         draw = ImageDraw.Draw(padded_image)
         font = ImageFont.load_default(size=int(pad_height / 2))
 
-        text = annotation.format(**cube.meta)
+        formatter = DefaultFormatter()
+        text = formatter.format(annotation, **cube.meta)
         text_offset = int(10 * image.shape[1] / 2048)
         text_position = (text_offset, pil_image.height + text_offset)
         draw.text(text_position, text, font=font, fill=(255, 255, 255))
@@ -163,7 +177,9 @@ def write_quicklook_to_mp4(files: list[str],
                            filename: str,
                            framerate: int = 5,
                            resolution: int = 1024,
-                           codec: str = "libx264") -> None:
+                           codec: str = "libx264",
+                           ffmpeg_cmd: str = "ffmpeg",
+                           ) -> None:
     """
     Write a list of input quicklook jpeg2000 files to an output mp4 animation.
 
@@ -180,12 +196,14 @@ def write_quicklook_to_mp4(files: list[str],
     codec : str, optional
         Codec to use for encoding. For GPU acceleration.
         "h264_videotoolbox" can be used on ARM Macs, "h264_nvenc" can be used on Intel machines.
+    ffmpeg_cmd : str
+        path to the ffmpeg executable
 
     """
     input_sequence = f"concat:{'|'.join(files)}"
 
     ffmpeg_command = [
-        "ffmpeg",
+        ffmpeg_cmd,
         "-framerate", str(framerate),
         "-i", input_sequence,
         "-vf", f"scale=-1:{resolution}",
@@ -210,6 +228,7 @@ def write_ndcube_to_fits(cube: NDCube,
             _update_statistics(cube)
 
         full_header = cube.meta.to_fits_header(wcs=cube.wcs)
+        full_header["FILENAME"] = filename
 
         hdu_data = fits.CompImageHDU(data=cube.data,
                                      header=full_header,