+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 
+- Allows custom path for ffmpeg in quicklook movies in https://github.com/punch-mission/punchbowl/pull/438
 - Allows L1 calibration files to be passed in as callables in https://github.com/punch-mission/punchbowl/pull/426
 - Speedups to L1 production in https://github.com/punch-mission/punchbowl/pull/426
 - Adds metadata to output jpeg2000 files in https://github.com/punch-mission/punchbowl/pull/433